Break files apart to fix circular deps
Also added a few more unit tests & updated rollup minifier plugin
