+++ lib/a1.js
@@ -1,66 +1,45 @@
 import { MAX_ROWS, MAX_COLS } from './constants.js';
 import { parseRef } from './parseRef.js';
+import { toCol } from './toCol.js';
+import { fromCol } from './fromCol.js';
 import { stringifyPrefix, stringifyPrefixAlt } from './stringifyPrefix.js';
 
 const clamp = (min, val, max) => Math.min(Math.max(val, min), max);
 const toColStr = (c, a) => (a ? '$' : '') + toCol(c);
 const toRowStr = (r, a) => (a ? '$' : '') + toRow(r);
-const charFrom = String.fromCharCode;
 
-/**
- * Convert a column string representation to a 0 based
- * offset number (`"C"` = `2`).
- *
- * The method expects a valid column identifier made up of _only_
- * A-Z letters, which may be either upper or lower case. Other input will
- * return garbage.
- *
- * @param {string} columnString  The column string identifier
- * @returns {number}  Zero based column index number
- */
-export function fromCol (columnString) {
-  const x = (columnString || '');
-  const l = x.length;
-  let n = 0;
-  if (l > 2) {
-    const c = x.charCodeAt(l - 3);
-    const a = c > 95 ? 32 : 0;
-    n += (1 + c - a - 65) * 676;
-  }
-  if (l > 1) {
-    const c = x.charCodeAt(l - 2);
-    const a = c > 95 ? 32 : 0;
-    n += (1 + c - a - 65) * 26;
-  }
-  if (l) {
-    const c = x.charCodeAt(l - 1);
-    const a = c > 95 ? 32 : 0;
-    n += (c - a) - 65;
-  }
-  return n;
-}
-
-/**
- * Convert a 0 based offset number to a column string
- * representation (`2` = `"C"`).
- *
- * The method expects a number between 0 and 16383. Other input will
- * return garbage.
- *
- * @param {number} columnIndex Zero based column index number
- * @returns {string} The column string identifier
- */
-export function toCol (columnIndex) {
-  return (
-    (columnIndex >= 702
-      ? charFrom(((((columnIndex - 702) / 676) - 0) % 26) + 65)
-      : '') +
-    (columnIndex >= 26
-      ? charFrom(((((columnIndex / 26) - 1) % 26) + 65))
-      : '') +
-    charFrom((columnIndex % 26) + 65)
-  );
-}
+// /**
+//  * Convert a column string representation to a 0 based
+//  * offset number (`"C"` = `2`).
+//  *
+//  * The method expects a valid column identifier made up of _only_
+//  * A-Z letters, which may be either upper or lower case. Other input will
+//  * return garbage.
+//  *
+//  * @param {string} columnString  The column string identifier
+//  * @returns {number}  Zero based column index number
+//  */
+// export function fromCol (columnString) {
+//   const x = (columnString || '');
+//   const l = x.length;
+//   let n = 0;
+//   if (l > 2) {
+//     const c = x.charCodeAt(l - 3);
+//     const a = c > 95 ? 32 : 0;
+//     n += (1 + c - a - 65) * 676;
+//   }
+//   if (l > 1) {
+//     const c = x.charCodeAt(l - 2);
+//     const a = c > 95 ? 32 : 0;
+//     n += (1 + c - a - 65) * 26;
+//   }
+//   if (l) {
+//     const c = x.charCodeAt(l - 1);
+//     const a = c > 95 ? 32 : 0;
+//     n += (c - a) - 65;
+//   }
+//   return n;
+// }
 
 export function fromRow (rowStr) {
   return +rowStr - 1;
