+++ lib/a1.spec.js
@@ -1,8 +1,6 @@
 /* eslint-disable object-property-newline, object-curly-newline */
 import { test, Test } from 'tape';
 import {
-  fromCol,
-  toCol,
   fromRow,
   toRow,
   toRelative,
@@ -33,15 +31,6 @@ Test.prototype.isA1Equal = function isA1Equal (expr, expect, opts) {
 
 // What happens when B2:A1 -> should work!
 test('convert to and from column and row ids', t => {
-  t.is(fromCol('a'), 0);
-  t.is(fromCol('A'), 0);
-  t.is(fromCol('AA'), 26);
-  t.is(fromCol('zz'), 701);
-  t.is(fromCol('ZZZ'), 18277);
-  t.is(toCol(0), 'A');
-  t.is(toCol(26), 'AA');
-  t.is(toCol(701), 'ZZ');
-  t.is(toCol(18277), 'ZZZ');
   t.is(fromRow('11'), 10);
   t.is(fromRow('1'), 0);
   t.is(toRow(12), '13');
