+++ lib/addTokenMeta.js
@@ -1,6 +1,6 @@
 import { REF_RANGE, REF_BEAM, REF_TERNARY, UNKNOWN, REF_STRUCT } from './constants.js';
 import { parseA1Ref } from './a1.js';
-import { parseStructRef } from './sr.js';
+import { parseStructRef } from './parseStructRef.js';
 
 function getIDer () {
   let i = 1;
