+++ lib/fixRanges.js
@@ -1,6 +1,7 @@
 import { isRange } from './isType.js';
 import { parseA1Ref, stringifyA1Ref, addA1RangeBounds } from './a1.js';
-import { parseStructRef, stringifyStructRef } from './sr.js';
+import { parseStructRef } from './parseStructRef.js';
+import { stringifyStructRef } from './stringifyStructRef.js';
 import { tokenize } from './lexer.js';
 import { REF_STRUCT } from './constants.js';
