+++ lib/index.js
@@ -16,9 +16,10 @@ export {
   isWhitespace
 } from './isType.js';
 
+export { fromCol } from './fromCol.js';
+export { toCol } from './toCol.js';
+
 export {
-  fromCol,
-  toCol,
   parseA1Ref,
   stringifyA1Ref,
   addA1RangeBounds
@@ -29,10 +30,8 @@ export {
   stringifyR1C1Ref
 } from './rc.js';
 
-export {
-  parseStructRef,
-  stringifyStructRef
-} from './sr.js';
+export { stringifyStructRef } from './stringifyStructRef.js';
+export { parseStructRef } from './parseStructRef.js';
 
 import {
   // token types
