+++ lib/lexerParts.js
@@ -18,8 +18,8 @@ import {
   MAX_ROWS,
   OPERATOR_TRIM
 } from './constants.js';
-import { fromCol } from './a1.js';
-import { parseSRange } from './sr.js';
+import { fromCol } from './fromCol.js';
+import { parseSRange } from './parseSRange.js';
 
 const re_ERROR = /^#(NAME\?|FIELD!|CALC!|VALUE!|REF!|DIV\/0!|NULL!|NUM!|N\/A|GETTING_DATA\b|SPILL!|UNKNOWN!|FIELD\b|CALC\b|SYNTAX\?|ERROR!|CONNECT!|BLOCKED!|EXTERNAL!)/i;
 const re_OPERATOR = /^(<=|>=|<>|[-+/*^%&<>=]|[{},;]|[()]|@|:|!|#)/;
