+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@borgar/fx",
-  "version": "4.10.2",
+  "version": "4.11.0",
   "description": "Utilities for working with Excel formulas",
   "main": "dist/fx.js",
   "types": "dist/fx.d.ts",
@@ -48,12 +48,12 @@
     "@borgar/eslint-config": "~3.1.0",
     "@borgar/jsdoc-tsmd": "~0.2.1",
     "@rollup/plugin-babel": "~6.0.4",
+    "@rollup/plugin-terser": "~0.4.4",
     "babel-eslint": "~10.1.0",
     "eslint": "~8.56.0",
     "eslint-plugin-jsdoc": "~48.1.0",
     "jsdoc": "~4.0.2",
     "rollup": "~4.12.0",
-    "rollup-plugin-minification": "~0.2.0",
     "tap-min": "~3.0.0",
     "tape": "~5.7.5",
     "typescript": "~5.3.3"
